+++ app/src/main/java/com/vad/ltale/presentation/PostViewModel.kt
@@ -17,11 +17,11 @@ class PostViewModel(private val postRepository: PostRepository) : ViewModel() {
     var posts: MutableLiveData<List<PostResponse>> = MutableLiveData()
     val post: MutableLiveData<Int> = MutableLiveData()
 
-    fun getPostsByUserId(userId: Int) = viewModelScope.launch {
+    fun getPostsByUserId(userId: Long) = viewModelScope.launch {
         posts.postValue(postRepository.getPostByUserId(userId))
     }
 
-    fun getPostById(postId: Int, position: Int) = viewModelScope.launch {
+    fun getPostById(postId: Long, position: Int) = viewModelScope.launch {
         postRepository.getPostById(postId)
         post.postValue(position)
     }
